platform: ext: target: stm: stm32wba65ri-dk2: Add flash gap for OEMxROT
Add a gap in the flash layout stm32wba65ri-dk1 target to leave room
for an external bootloader as OEMiROT that currently requires 96kB
at the beginning of the flash memory.

The gap may need to be enlarge, depending on the bootloader to
be embedded, as for example if enabling OEMiROT swap method for
updates or embedding OEMuROT together with OEMiROT. For this, a more
flexible way would be to have a configuration directive like
FLASH_LAYOUT_BEGIN_OFFSET to define this offset value. It could be
applicable to all STM32 targets of TF-M.

Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>