Platform: STM32WBA65I_DK: fix secure boot address in regression.sh
Correct the value used to program SECBOOTADD0 in regression.sh script
that may not be 0x180488.

The correct value to use should be set by script shell variable
secbootadd0 that is set thanks to postbuild.sh script according to the
flash layout used to build TF-M firmware. Add need to add the TF-M
header (1kByte) since it is not considered by postbuild script.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>